Fix/new assets page (#335)
* fix: fix

* fix: fix page loading end switch menu nav
