+++ src/components/pages/assets/AssetsInput.tsx
