+++ src/components/ramp/QuotationOffRamp.tsx
@@ -13,7 +13,7 @@ import { PiArrowDown, PiArrowRight } from 'react-icons/pi'
 import styled from 'styled-components'
 import { useDebounce } from 'usehooks-ts'
 import { useAccount } from 'wagmi'
-import AssetInput from '../assets/AssetsInput'
+import AssetInput from '../pages/assets/AssetsInput'
 import { TokenBalance } from '@/hooks/contracts/useBalanceOf'
 import SkeletonLoading from '../shared/icons/SkeletonLoading'
 import AlertMessageComponent from '../shared/AlertMessageComponent'
